\name{ri-package}
\alias{ri-package}
\alias{ri}
\docType{package}
\title{
ri: R package for performing randomization-based inference for experiments
}
\description{
This package provides a set of tools for conducting exact or approximate randomization-based inference for experiments of arbitrary design. The primary functionality of the package is in the generation, manipulation and use of permutation matrices implied by given experimental designs. Among other features, the package facilitates estimation of average treatment effects, constant effects variance estimation, randomization inference for significance testing against sharp null hypotheses and visualization of data and results.
}
\details{
\tabular{ll}{
Package: \tab ri\cr
Type: \tab Package\cr
Version: \tab 0.9\cr
Date: \tab 2012-05-10\cr
License: \tab GPL (>= 2)\cr
}
This package provides a set of tools for conducting exact or approximate inference for randomized experiments of arbitrary design. The primary functionality of the package is in the generation, manipulation and use of permutation matrices implied by given experimental designs. Among other features, the package facilitates estimation of average treatment effects, constant effects variance estimation and randomization inference for significance testing against sharp null hypotheses.
}
\author{
Peter M. Aronow <peter.aronow@yale.edu> and Cyrus Samii <cds2083@nyu.edu>

Maintainer: Cyrus Samii <cds2083@nyu.edu> 
}
\references{
Gerber, Alan S. and Donald P. Green. 2012. \emph{Field Experiments: Design, Analysis, and Interpretation}. New York: W.W. Norton.
}
